🔨 drop large hit, add small hit component
