+++ site/owid.scss
@@ -59,7 +59,6 @@
 @import "./search/SearchTopicsRefinementList.scss";
 @import "./search/_DEPRECATEDChartHit.scss";
 @import "./search/SearchChartHitMedium.scss";
-@import "./search/SearchChartHitLarge.scss";
 @import "./search/SearchActiveFilters.scss";
 @import "./search/SearchFilterPill.scss";
 @import "./search/SearchAutocompleteItemContents.scss";
